+++ CHANGELOG.md
@@ -70,6 +70,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `jj config get` now supports displaying array and table config values.
 
+* `jj util exec` sets the environment variable `JJ_WORKSPACE_ROOT`
+
 ### Fixed bugs
 
 * Fetching repositories that have submodules no longer errors even if